requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
allure-python-commons==2.13.2
async-timeout==5.0.1
attrs==23.1.0
backports.asyncio.runner==1.2.0
boto3==1.26.137
botocore==1.29.137
certifi==2023.5.7
chardet==5.2.0
charset-normalizer==3.1.0
colorama==0.4.6
exceptiongroup==1.3.0
frozenlist==1.7.0
humanize==4.12.3
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.21.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
multidict==6.6.3
narwhals==1.42.0
numpy==2.0.2
packaging==23.1
pandas==2.3.0
pillow==11.2.1
plotly==6.1.2
pluggy==1.6.0
propcache==0.3.2
py==1.11.0
Pygments==2.15.1
pytest-asyncio==1.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.1
requests==2.28.2
rich==13.3.5
s3transfer==0.6.2
six==1.17.0
tabulate==0.9.0
tomli==2.2.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==1.26.18
yarl==1.20.1
